GD to HoaDon

+++ app/Http/Controllers/admin/KhoaHocController.php
@@ -7,6 +7,7 @@ use App\Models\BaiHoc;
 use App\Models\DanhMuc;
 use App\Models\KhoaHoc;
 use App\Models\DanhMucCon;
+use App\Models\HoaDon;
 use App\Models\LopHoc;
 use App\Models\LopHoc_BaiHoc;
 use App\Models\TaiKhoan;
@@ -172,9 +173,12 @@ class KhoaHocController extends Controller
     public function delete($id)
     {
         $khoahoc = KhoaHoc::find($id);
-        $khoahoc->delete();
-
-        return redirect('admin/khoahoc/')->with('thongbao', 'Xóa thành công!');
+        $hoadon = HoaDon::where('MAKH', $id)->get();
+        if (!$hoadon) {
+            $khoahoc->delete();
+            return redirect('admin/khoahoc/')->with('thongbao', 'Xóa thành công!');
+        } else
+            return redirect('admin/khoahoc/')->with('thongbao', 'Khóa học đã có người mua.\n Không thể xóa khóa học này!');
     }
 
     public function search(Request $request)
